requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
base58==2.1.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
ed25519==1.5
idna==3.3
psycopg2==2.9.2
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
